pyproject.toml
[project]
name = "adversarial-redteam"
version = "0.0.1"
description = "Adversarial-distribution red team for wargaming. SCSP Hackathon 2026."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    # LLM access. Direct SDKs, no LiteLLM (per RA-6 §"LiteLLM unified layer":
    # for two providers at hackathon scale, the abstraction cost outweighs the benefit).
    "anthropic>=0.39",
    "openai>=1.55",
    "tenacity>=9.0",

    # Doctrine corpus is markdown + YAML frontmatter (see PROJECT_SPEC.md §5).
    # No Chroma / no chunker. sentence-transformers stays for agent-memory embeddings.
    "sentence-transformers>=3.3",

    # Schemas, config
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",

    # Numerics
    "numpy>=1.26",
    "scikit-learn>=1.5",  # KMeans for convergence clustering

    # UI
    "streamlit>=1.40",
    "plotly>=5.24",       # cluster viz

    # Stdlib helpers
    "tomli-w>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]  # line length handled by formatter when we want
